Move Producer functions to AAW
The consumer and producer classes were solving the same problem with different styles at different levels of efficiency.
I decided that it would be better to stick to DRY and combine the logic into one class, AAW.
At this point there is little reason to have AAW, Consumer, and Producer classes when there is little
benefit from having each of the classes. Right now the Consumer and Producer are only subclasses of AAW, with the Producer having an extra parameter.