add predecessor method, field & init param to AAW
The predecessor is a convenience member that makes queueing to proceeding and preceding workers easier.
Without it, one would have to remember the names of the other workers when accessing their queue.
This becomes troublesome when renaming a worker as the refactoring would have to be done manually

Other changes:
- make each AAW have their own stats Counter
- change tui to reflect stats changes
- change StatDisplay to reflect stats changes
- renamed AAW to Job
- renamed loopmanager to jobmanager
- add maxsize to queuemanager
- more visual improvements to stats