job: fix #queue_finished() bug v13.0.0
Before when adding new items to the queue after #queue_finished was called, the new added items would not be processed until later.
This is because I could not find a solution that would allow the program to process the new items and still close successfully.
I've fixed this issue by adding a queue watcher loop that will keep track of the queue and call queue_finished() when all successors and predecessors are finished processing and sending their items.